Moved bool_button to funcs.rs and generalised it
